% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paf_utils.R
\name{apply_modifications}
\alias{apply_modifications}
\title{Modify data frame for PAF calculations}
\usage{
apply_modifications(df, modlist)
}
\arguments{
\item{df}{a data frame to modify}

\item{modlist}{named list of vectors or scalars to convert to}
}
\value{
a modified data frame
}
\description{
This function returns a data frame which has had the specified modifications
applied.
}
\details{
\code{modlist} should be a named list of scalars or vectors. Each name should
correspond to a column in \code{df}. The first element in each vector should
be the desired output value. If there are subsequent elements in the vector,
these identify which values to convert from.
}
\section{Warning}{
 Column types (particularly of factors) are prone to change,
  and might drop levels. Hence using the \code{xlev} argument of
  \code{model.frame} is wise.
}

\examples{
# changes smoking status to 'Never' for smokers who haven't already quit.
apply_modifications(minifhs, list(SMOKE = c("Never", "<30/day", ">=30/day")))
## (note no change to rows with value "Former")

# changes BMI status from non-overweight, regardless of previous value
( beware <- apply_modifications(minifhs, list(BMI_2  = "<25.0")) )

# beware!
str(minifhs$BMI_2)
str(beware$BMI_2)

}
\keyword{internal}
