% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data_fun.R
\name{collapse_times}
\alias{collapse_times}
\title{Collapse time-to-event data from multiple events}
\usage{
collapse_times(indata, ind1, time1, ..., time.out = "time")
}
\arguments{
\item{indata}{data frame to be modified}

\item{ind1, time1, ...}{columns representing indicator and time data
(respectively) for each event, repeated for each event}

\item{time.out}{name of the output time column}
}
\value{
a data frame with a single column of event times, with modified
  indicators such that only the first events are recorded
}
\description{
This function converts multiple time-to-event columns into one time column,
with indicator columns that are \code{TRUE} only for the first event (or
first events, if these occur simultaneously).
}
\examples{
# The original data frame we will work with has two time
# columns -- one for diabetes incidence (DIAB_FT), and one
# for death (DEATH_FT).
head(minifhs)

# we are interested in the first of diabetes incidence and death
ct_minifhs <- collapse_times(minifhs, "DIAB", "DIAB_FT", "DEATH", "DEATH_FT")
head(col_minifhs)
}
