% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paf_utils.R
\name{design_frames}
\alias{design_frames}
\title{Create design frames from data}
\usage{
design_frames(df, terms, modifications, xlev)
}
\arguments{
\item{df}{initial data frame}

\item{terms}{object of terms with response dropped}

\item{modifications}{list of modifications to make (see
\code{apply_modifications} for format)}

\item{xlev}{xlevels object from the survival regression}
}
\value{
a list of two model frames, \code{design} and \code{modified},
  carrying predictor values
}
\description{
Create design frames from data
}
\keyword{internal}
