% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_calc.R
\name{dpaf_delta_Sp}
\alias{dpaf_delta_Sp}
\title{Calculate survival differences for dpaf study}
\usage{
dpaf_delta_Sp(Sp, ID, PERIOD)
}
\arguments{
\item{Sp}{vector of survival products}

\item{ID}{vector of corresponding IDs}

\item{PERIOD}{vector of corresponding periods}
}
\value{
vector of survival product differences over the course of each period
}
\description{
Assumes arguments are sorted -- check with \code{\link{is_period_unsorted}}.
}
\keyword{internal}
