% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_calc.R
\name{dpaf_gipaf}
\alias{dpaf_gipaf}
\title{Calculate gradients of \eqn{\log(1-PAF)} wrt disease and mortality
coefficients}
\usage{
dpaf_gipaf(gi_mod, i_mod, gi_raw, i_raw)
}
\arguments{
\item{gi_mod}{named list of gradients of \eqn{I^*}}

\item{i_mod}{\eqn{I^*}}

\item{gi_raw}{named list of gradients of \eqn{I}}

\item{i_raw}{\eqn{I}}
}
\value{
a named list of gradient vectors of the iPAF
}
\description{
Calculate gradients of \eqn{\log(1-PAF)} wrt disease and mortality
coefficients
}
\keyword{internal}
