% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_calc.R
\name{dpaf_grad_delta_Sp}
\alias{dpaf_grad_delta_Sp}
\title{Calculate gradient of survival products}
\usage{
dpaf_grad_delta_Sp(grad_S_l, Sp, ID, PERIOD)
}
\arguments{
\item{grad_S_l}{named list of survival gradient matrices}

\item{Sp}{vector of survival products}

\item{ID}{corresponding vector of IDs}

\item{PERIOD}{corresponding vector of PERIODs}
}
\value{
named list of survival product gradient matrices
}
\description{
Assumes arguments are sorted -- check with \code{\link{is_period_unsorted}}.
}
\keyword{internal}
