% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_calc.R
\name{dpaf_grad_I}
\alias{dpaf_grad_I}
\title{Calculate average/expected disease incidence \eqn{I} for dpaf study}
\usage{
dpaf_grad_I(grad_lambda_l, grad_dSp_l, lambda_l, dSp, PERIOD)
}
\arguments{
\item{grad_lambda_l}{named list of hazard gradient matrices}

\item{lambda_l}{named list of hazards}

\item{dSp}{vector of survival differences}

\item{PERIOD}{vector of corresponding periods}

\item{grad_dSp}{named list of survival product difference matrices}

\item{ID}{vector of corresponding IDs}
}
\value{
named vector of mortalities \eqn{I} (see Warning on names)
}
\description{
Calculate average/expected disease incidence \eqn{I} for dpaf study
}
\keyword{internal}
