% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_calc.R
\name{dpaf_grad_lambda}
\alias{dpaf_grad_lambda}
\title{Calculate hazard gradients for dpaf study}
\usage{
dpaf_grad_lambda(z, cf_l)
}
\arguments{
\item{z}{model matrix}

\item{cf_l}{named list of vectors \eqn{\gamma^D, \gamma^M} of parameter
estimates}
}
\value{
named list of hazard gradients (in matrix form)
}
\description{
Calculate hazard gradients for dpaf study
}
\keyword{internal}
