% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_calc.R
\name{dpaf_grad_S}
\alias{dpaf_grad_S}
\title{Calculate gradient of survivals for mpaf study}
\usage{
dpaf_grad_S(grad_lambda_l, S_l, ID, PERIOD, dt)
}
\arguments{
\item{grad_lambda_l}{named list of matrices of hazard gradients}

\item{S_l}{named list of vector of survivals}

\item{ID}{corresponding vector of IDs}

\item{PERIOD}{corresponding vector of periods}

\item{dt}{delta-times -- length of each period}
}
\value{
names list of matrix of gradients, where each row in each list item
  is the gradient of the corresponding survival
}
\description{
Calculate gradient of survivals for mpaf study
}
\keyword{internal}
