% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_calc.R
\name{dpaf_I}
\alias{dpaf_I}
\title{Calculate average/expected disease incidence \eqn{I} for dpaf study}
\usage{
dpaf_I(lambda_l, dSp, PERIOD)
}
\arguments{
\item{lambda_l}{named list of hazards}

\item{dSp}{vector of survival differences -- if expected disease incidence
over a period \eqn{(a_{j-1}, a_j]} is desired, then \code{dSp} should be
\eqn{\Delta{S}} as output by \code{\link{dpaf_delta_Sp}}. If expected
disease incidence over a period \eqn{(0, a_j]} is desired, then calculate
\eqn{I} as before and then take the sum.}

\item{PERIOD}{vector of corresponding periods}

\item{ID}{vector of corresponding IDs}
}
\value{
named vector of mortalities \eqn{I} (see Warning on names)
}
\description{
Calculate average/expected disease incidence \eqn{I} for dpaf study
}
\section{Warning}{
 the names of the vector are exactly the levels of
  \code{PERIOD}, however this is a misnomer if calculating cumulative
  mortalities.
}

\keyword{internal}
