% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_calc.R
\name{dpaf_lambda}
\alias{dpaf_lambda}
\title{Calculate hazards for dpaf study}
\usage{
dpaf_lambda(z, cf_l)
}
\arguments{
\item{z}{model matrix}

\item{cf_l}{named list of vectors \eqn{\gamma^D, \gamma^M} of parameter
estimates}
}
\value{
named list of hazards
}
\description{
Calculate hazards for dpaf study
}
\keyword{internal}
