% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_calc.R
\name{dpaf_S}
\alias{dpaf_S}
\title{Calculate survivals for mpaf study}
\usage{
dpaf_S(lambda_l, ID, PERIOD, dt)
}
\arguments{
\item{lambda_l}{named list of vectors of hazards corresponding to \code{ID}s
and \code{PERIOD}s}

\item{ID}{vector of IDs}

\item{PERIOD}{vector of periods}

\item{dt}{delta-times -- length of each period}
}
\value{
named list of survivals
}
\description{
Calculate survivals for mpaf study
}
\keyword{internal}
