% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_calc.R
\name{dpaf_Sp}
\alias{dpaf_Sp}
\title{Calculate disease-free survival (**S**urvival **p**roduct)}
\usage{
dpaf_Sp(S_l)
}
\arguments{
\item{S_l}{names list of survivals \eqn{S^D, S^M}}
}
\value{
vector of survival products \eqn{S = S^D S^M}
}
\description{
Calculate disease-free survival (**S**urvival **p**roduct)
}
\keyword{internal}
