% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_paf.R
\name{dpaf_summary}
\alias{dpaf_summary}
\title{Calculate mortality PAF estimates, confidence intervals and groupwise
differences}
\usage{
dpaf_summary(disease_resp, death_resp, predictors, dpaf_data,
  modifications, covar_model, prevalence_data, group_vars, level = 0.95,
  ...)
}
\arguments{
\item{disease_resp}{formula with response for disease regression (note RHS
should be \code{.})}

\item{death_resp}{formula with response for mortality regression (note RHS
should be \code{.})}

\item{predictors}{formula with RHS containing predictors for both survival
regressions (LHS should be empty)}

\item{dpaf_data}{an object of class \code{dpaf_response}, from
\code{\link{gen_data}}}

\item{modifications}{a list of modifications to apply for PAF calculation;
see Details for more information}

\item{covar_model}{an optional character vector describing variables of
interest for hazard ratio calculation}

\item{prevalence_data}{optional matrix of external prevalences as class
\code{paf_data}}

\item{group_vars}{names of columns to group by in PAF difference estimations}

\item{level}{width of confidence intervals, default 0.95}

\item{...}{extra parameters to be passed to \code{\link{est_matrix}}}
}
\value{
a list of potentially interesting hazard ratio and PAF estimates
}
\description{
Calculate mortality PAF estimates, confidence intervals and groupwise
differences
}
\details{
If \code{covar_model} is not specified,
the "hazard ratios" reported include all first order terms.

\code{modifications} should be a named list of scalars or vectors. Each name
should correspond to a column in \code{df}. The first element in each vector
should be the desired output value. If there are subsequent elements in the
vector, these identify which values to convert from.
}
