% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_matrix.R
\name{est_matrix}
\alias{est_matrix}
\title{Calculate model frames, parameter estimates and hazard ratios}
\usage{
est_matrix(sr_formula, paf_response, modifications, hr_out, level = 0.95,
  ...)
}
\arguments{
\item{sr_formula}{a formula to be passed to \code{\link[survival]{survreg}}}

\item{paf_response}{a response object as output by \code{\link{gen_data}}}

\item{modifications}{a list of modifications to apply for PAF calculation;
see Details for more information}

\item{hr_out}{an optional character vector describing variables of
interest for hazard ratio calculation}

\item{level}{width of the confidence intervals for hazard ratios, default
0.95.}

\item{...}{extra parameters to be passed to \code{survreg}}
}
\value{
a list of type \code{paf_est_matrix} with the following items:

  \item{est_matrix_call}{the call to this function}

  \item{modifications}{the list of modifications applied}

  \item{survreg}{the survival regression object}

  \item{coefficients}{the parameter estimates as in \code{survreg}}

  \item{var}{the covariance matrix as in \code{survreg}}

  \item{terms}{a \code{terms} object for the survival regression with
  responses removed}

  \item{HR}{the hazard ratios and their confidence intervals; see Warning}

  \item{design}{the model frame of predictors}

  \item{modified}{the model frame of predictors after modification}
}
\description{
This function fits a survival regression and returns model frames (without
response columns) with and without modifications applied, parameter estimates
and their standard errors, and hazard ratios with their confidence intervals.
Note that hazard ratios reported as \code{NA} are reference levels.
}
\details{
\code{modifications} should be a named list of scalars or
vectors. Each name should correspond to a column in \code{df}. The
first element in each vector should be the desired output value. If
there are subsequent elements in the vector, these identify which
values to convert from.

If \code{hr_out} is not specified,
the "hazard ratios" reported include all first order terms.
}
\section{Warnings}{


  A hazard ratio is not well-defined when the term of interest is an
  interaction. Use caution when interpreting such results.

  When using interaction terms in \code{hr_out}, the wrong order of
  covariate names will result in an error. Only terms that appear in
  \code{attr(terms(sr_formula), "term.labels")} can be used.
}

