% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data_fun.R
\name{expand_events}
\alias{expand_events}
\title{Expand time-to-event data from multiple events}
\usage{
expand_events(indata, event, prefix = paste0(event, "."))
}
\arguments{
\item{indata}{data frame to be modified}

\item{event}{column of factor describing the event that occurred}

\item{prefix}{prefix to append to new column names}
}
\value{
a data frame with the original event column removed, and replaced
  with an indicator column for each level of the event factor
}
\description{
This function converts an event factor from time-to-event data into multiple
indicator columns. The indicator columns are \code{TRUE} if the corresponding
event was recorded.
}
\examples{
# Our original data frame with liver transplant data has one
# of four events occurring, stored in the "event" column.
head(survival::transplant)

# We want to convert the event column into multiple indicator
# columns
tsplt_ind <- expand_events(survival::transplant, "event")
head(tsplt_ind)
}
