% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_models}
\alias{fit_models}
\title{Fit all survival regression models for PAF data}
\usage{
fit_models(predictors, paf_data, ...)
}
\arguments{
\item{predictors}{a formula with empty LHS containing the predictors of interest}

\item{paf_data}{an extended data frame as from \code{\link{gen_data_fun}}}

\item{...}{further objects to be passed to \code{\link[survival]{survreg}}}
}
\value{
a list of \code{survreg} objects from the \code{survival} package
}
\description{
This function fits a survival regression object for every event specified in
a PAF data frame. It returns a list object with names determined by the
column names of the events. The primary event will always occur first in the
list.
}
\examples{
# The original data frame we will work with has two time
# columns -- one for diabetes incidence (DIAB_FT), and one
# for death (DEATH_FT).
head(minifhs)

# we are interested in the first of diabetes incidence and death
ct_minifhs <- collapse_times(minifhs, "DIAB", "DIAB_FT", "DEATH", "DEATH_FT")
head(ct_minifhs)

# Now we can time-vary the data
ext_minifhs <- gen_data_fun(ct_minifhs, c(0,5,10,15,17), time = "time",
                            primary = "DIAB", secondary = "DEATH",
                            id_col = "ID")

# Fit survival regression models for every event, and then summarise
mods_minifhs <- fit_models(~ B_COHORT * period + SEX + BMI_2, ext_minifhs)
mods_minifhs
}
