% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{gen_data}
\alias{gen_data}
\alias{mpaf_gen_data}
\title{Generate mortality PAF data from cohort data.}
\usage{
gen_data(indata, id_var, ft_breaks, death_ind, death_time, disease_ind,
  disease_time, variables = character(0),
  na.action = getOption("na.action"), period_factor = "f_period",
  time_var = "f_end")

mpaf_gen_data(indata, id_var, ft_breaks, death_ind, death_time,
  variables = character(0), na.action = getOption("na.action"),
  period_factor = "f_period", time_var = "f_end")
}
\arguments{
\item{indata}{Original data frame containing data from cohort study}

\item{id_var}{Name of column containing individual IDs}

\item{ft_breaks}{Numeric vector of breaks between time periods}

\item{death_ind}{(Optional) Name of logical column indicating if death
occurred before censoring (i.e. \code{TRUE} if an individual died before
the end of follow-up)}

\item{death_time}{(Optional) Name of numerical column with individual time
until death or censoring}

\item{disease_ind}{(Optional) Name of logical column indicating if disease
occurred before censoring}

\item{disease_time}{(Optional) Name of numerical column with individual time
until disease or censoring}

\item{variables}{Character vector of predictor columns to keep from
\code{indata}}

\item{na.action}{Function (found via call to \code{match.fun}) which
indicates how this function should treat missing data}

\item{period_factor, time_var}{Names of new columns in data frame representing
individual follow-up period and follow-up time respectively}
}
\value{
A list of class \code{"paf_data"} containing these elements:

  \item{data_call}{The function call to generate data in this form}

  \item{nobs}{The number of observations used from the initial data frame}

  \item{na.action}{The IDs removed from the data due to missingness}

  \item{breaks}{A vector of times where baseline hazard changes}

  \item{ID}{A vector of IDs corresponding to rows of the data frame}

  \item{PERIOD}{A factor vector of periods corresponding to rows of the data
  frame}

  \item{data}{A data frame containing predictor (and possibly response)
  columns as given in the \code{variables} argument, as well as a column for
  the period (if requested)}

  If a response is included, the returned list also inherits
  \code{"dpaf_response"} or \code{"mpaf_response"} for disease or mortality
  PAF calculations respectively.
}
\description{
Generate mortality PAF data from cohort data.
}
