% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data_fun.R
\name{gen_data_fun}
\alias{gen_data_fun}
\title{Generate time-varying cohort data}
\usage{
gen_data_fun(indata, ft_breaks, time, primary, secondary = NULL, id_col,
  period_factor = "period")
}
\arguments{
\item{indata}{data frame to be extended in time}

\item{ft_breaks}{numeric vector of breaks between time periods}

\item{time}{name of response column for time}

\item{primary}{name of response column for event of interest}

\item{secondary}{names of response columns for competing risk events}

\item{id_col}{optional name of column containing unique identifiers for each
row}

\item{period_factor}{name of new column containing time period}
}
\value{
a data frame with the at least the \code{ft_breaks}, \code{time},
  \code{primary}, \code{secondary} and \code{period_factor} arguments as
  attributes. Further, the data frame has an \code{id} attribute that gives
  unique values for each individual, which may have been constructed if
  \code{id_col} was not given.

  The class of the data frame has been prepended with \code{paf_data} if
  there are no competing events, or \code{cr_paf_data} if potential competing
  risks hav ebeen included.
}
\description{
This function takes time-to-event data and converts it to a format that
varies with time. The original data consist of a data frame with covariate
columns and response columns. The response columns should include one
numerical column for time, plus one logical column for each event indicating
if this event was the first to occur.
}
\examples{
# The original data frame we will work with has two time
# columns -- one for diabetes incidence (DIAB_FT), and one
# for death (DEATH_FT).
head(minifhs)

# we are interested in the first of diabetes incidence and death
ct_minifhs <- collapse_times(minifhs, "DIAB", "DIAB_FT", "DEATH", "DEATH_FT")
head(ct_minifhs)

# Now we can time-vary the data
ext_minifhs <- gen_data_fun(ct_minifhs, c(0,5,10,15,17), time = "time",
                            primary = "DIAB", secondary = "DEATH",
                            id_col = "ID")

# Fit survival regression models for every event, and then summarise
mods_minifhs <- fit_models(~ B_COHORT * period + SEX + BMI_2, ext_minifhs)
mods_minifhs
}
