% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_ratio.R
\name{hazard_ratios}
\alias{hazard_ratios}
\title{Calculate hazard ratios from a survival regression object}
\usage{
hazard_ratios(survreg, hr_out, level = 0.95)
}
\arguments{
\item{survreg}{a \code{\link[survival]{survreg}} object of the model to base
hazard ratios upon}

\item{hr_out}{a list of variables to report hazard ratios for}

\item{level}{the desired widtch of the confidence interval, default 0.95}
}
\value{
A named nested list of hazard ratio tables. The top level list
  separates different risk factors requested in \code{hr_out}. The nested
  lists distinguish groups across which hazard ratios may differ.
}
\description{
This function reports hazard ratios from a survival analysis where (possibly
piecewise) constant hazards are assumed. Where interaction terms are present,
the function automatically determines which groups to calculate hazard ratios
for. Note that approximate confidence intervals based off the asymptotic
normality of log-hazard ratios are reported.
}
