% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_ratio.R
\name{hr_group}
\alias{hr_group}
\title{Calculate hazard ratios within a particular group}
\usage{
hr_group(mm, cf, vv, level)
}
\arguments{
\item{mm}{model matrix fragment of interest}

\item{cf}{coefficient vector of survival regression object}

\item{vv}{variance matrix from survival regression object}

\item{level}{confidence level desired}
}
\value{
matrix of hazard ratios and confidence intervals
}
\description{
Calculate hazard ratios within a particular group
}
