% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_calc.R
\name{mpaf_delta_S}
\alias{mpaf_delta_S}
\title{Calculate survival differences for mpaf study}
\usage{
mpaf_delta_S(S, ID, PERIOD)
}
\arguments{
\item{S}{vector of survivals}

\item{ID}{vector of corresponding IDs}

\item{PERIOD}{vector of corresponding periods}
}
\value{
vector of survival differences over the course of each period
}
\description{
Calculate survival differences for mpaf study
}
\keyword{internal}
