% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_est.R
\name{mpaf_est_diff}
\alias{mpaf_est_diff}
\title{Estimate significance of groupwise differences in mortality PAFs}
\usage{
mpaf_est_diff(mpaf1, mpaf2, vv)
}
\arguments{
\item{mpaf1, mpaf2}{objects of class \code{mpaf} (from
\code{\link{mpaf_est_paf}}) to be compared}

\item{vv}{the covariance matrix of parameters}
}
\value{
a matrix containing the PAF differences, their standard errors, Z
  values and p values.
}
\description{
The elements of the output from this function should be able to be passed to
\code{\link[stats]{printCoefmat}}.
}
