% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_est.R
\name{mpaf_est_paf}
\alias{mpaf_est_paf}
\title{Estimate mortality PAFs}
\usage{
mpaf_est_paf(mpaf_fit, paf_data, newdata, level = 0.95)
}
\arguments{
\item{mpaf_fit}{an object of class \code{paf_est_matrix}}

\item{paf_data}{an object of class \code{paf_data}, used for the
\code{\link{est_matrix}} call}

\item{newdata}{new prevalences, as an object of class \code{paf_data}}

\item{level}{width of confidence interval, default 0.95}
}
\value{
a list containing the following elements:

  \item{paf, paf0}{the mortality PAFs for individual and cumulative periods,
  respectively, with their confidence intervals as specified by the
  \code{level} argument}

  \item{se_ipaf, se_ipaf0}{the standard errors for \eqn{\log(1-PAF)}, as
  above}

  \item{grad_paf, grad_paf0}{the gradients of the PAFs as above, for
  difference calculations}
}
\description{
Estimate mortality PAFs
}
