% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_calc.R
\name{mpaf_grad_delta_S}
\alias{mpaf_grad_delta_S}
\title{Calculate differences of gradient of survivals for mpaf study}
\usage{
mpaf_grad_delta_S(grad_S, ID, PERIOD)
}
\arguments{
\item{grad_S}{matrix of survival gradients}

\item{ID}{corresponding vector of IDs}

\item{PERIOD}{corresponding vector of periods}
}
\value{
matrix of gradient differences, where each row is difference for the
  corresponding survival gradient
}
\description{
Assumes arguments are sorted -- check with \code{\link{is_period_unsorted}}.
}
\keyword{internal}
