% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_calc.R
\name{mpaf_grad_I}
\alias{mpaf_grad_I}
\title{Calculate average mortality gradient \eqn{\nabla{I}} for mpaf study}
\usage{
mpaf_grad_I(grad_dS, ID, PERIOD)
}
\arguments{
\item{grad_dS}{matrix of gradient survival differences -- if cumulative
mortality over time $(0, a_j]$ is desired, then \code{dS} should be
\code{-grad_S}, but if mortality over a period $(a_{j-1}, a_j]$ is desired,
then \code{dS} should be the differences as output by
\code{\link{mpaf_grad_delta_S}}.}

\item{ID}{vector of corresponding IDs}

\item{PERIOD}{vector of corresponding periods}
}
\value{
named matrix of mortalities $I$ (see Warning on names)
}
\description{
Calculate average mortality gradient \eqn{\nabla{I}} for mpaf study
}
\section{Warning}{
 the names of the vector are exactly the levels of
  \code{PERIOD}, however this is a misnomer if calculating cumulative
  mortalities.
}

\keyword{internal}
