% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_calc.R
\name{mpaf_grad_lambda}
\alias{mpaf_grad_lambda}
\title{Calculate gradient of hazard for mpaf study}
\usage{
mpaf_grad_lambda(z, lambda)
}
\arguments{
\item{z}{model matrix}

\item{lambda}{vector of hazards}
}
\value{
matrix of gradients, where each row is a gradient for the
  corresponding hazard
}
\description{
Calculate gradient of hazard for mpaf study
}
\keyword{internal}
