% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_calc.R
\name{mpaf_grad_S}
\alias{mpaf_grad_S}
\title{Calculate gradient of survivals for mpaf study}
\usage{
mpaf_grad_S(grad_lambda, S, ID, PERIOD, dt)
}
\arguments{
\item{grad_lambda}{matrix of hazard gradients}

\item{S}{vector of survivals}

\item{ID}{corresponding vector of IDs}

\item{PERIOD}{corresponding vector of periods}

\item{dt}{delta-times -- length of each period}
}
\value{
matrix of gradients, where each row is a gradient for the
  corresponding survival
}
\description{
Assumes arguments are sorted -- check with \code{\link{is_period_unsorted}}.
}
\keyword{internal}
