% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_calc.R
\name{mpaf_I}
\alias{mpaf_I}
\title{Calculate average mortality $I$ for mpaf study}
\usage{
mpaf_I(dS, PERIOD)
}
\arguments{
\item{dS}{vector of survival differences -- if cumulative mortality over time
\eqn{(0, a_j]} is desired, then \code{dS} should be \code{1-S}, but if
mortality over a period \eqn{(a_{j-1}, a_j]} is desired, then \code{dS}
should be \eqn{\Delta{S}} as output by \code{\link{mpaf_delta_S}}.}

\item{PERIOD}{vector of corresponding periods}

\item{ID}{vector of corresponding IDs}
}
\value{
named vector of mortalities $I$ (see Warning on names)
}
\description{
Calculate average mortality $I$ for mpaf study
}
\section{Warning}{
 the names of the vector are exactly the levels of
  \code{PERIOD}, however this is a misnomer if calculating cumulative
  mortalities.
}

\keyword{internal}
