% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_calc.R
\name{mpaf_lambda}
\alias{mpaf_lambda}
\title{Calculate hazards for mpaf study}
\usage{
mpaf_lambda(z, cf)
}
\arguments{
\item{z}{model matrix}

\item{cf}{named vector \eqn{\gamma} of parameter estimates}
}
\value{
named matrix of hazards
}
\description{
Calculate hazards for mpaf study
}
\keyword{internal}
