% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paf_utils.R
\name{name_by}
\alias{name_by}
\title{Convert a \code{by} object to a named list}
\usage{
name_by(x, lsep = ": ", fsep = ", ")
}
\arguments{
\item{x}{a \code{by} object}

\item{lsep}{character separator between a factor name and the level}

\item{fsep}{character separator between two factors}
}
\value{
a named, flattened list
}
\description{
Convert a \code{by} object to a named list
}
\keyword{internal}
