% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paf_utils.R
\name{paf_data_split}
\alias{paf_data_split}
\title{Split PAF data into groups by a factor}
\usage{
paf_data_split(object, INDICES, lsep = ": ", fsep = ", ")
}
\arguments{
\item{object}{An object of class \code{paf_data} (from
\code{\link{gen_data}}.}

\item{INDICES}{A factor or list of factors that can be passed to
\code{\link{split}}. Can optionally be named if a list.}

\item{lsep}{character separator between a factor name and the level}

\item{fsep}{character separator between two factors}
}
\value{
A named list of objects with class \code{dpaf_data} or
  \code{paf_data}, with names corresponding to the names of the list output
  by \code{split}. If a named list was given in \code{INDICES}, then these
  names will be included.
}
\description{
This function is a wrapper around R-base's \code{split} function designed for
\code{paf_data} structures.
}
