% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_paf.R
\name{mpaf_summary}
\alias{mpaf_summary}
\title{Calculate mortality PAF estimates, confidence intervals and groupwise
differences}
\usage{
mpaf_summary(sr_formula, paf_data, modifications, hr_out, prevalence_data,
  group_vars, level = 0.95, ...)
}
\arguments{
\item{sr_formula}{a formula to be passed to \code{\link[survival]{survreg}}}

\item{paf_data}{an object containing the response as in \code{\link{gen_data}}}

\item{modifications}{a list of modifications to apply for PAF calculation;
see Details for more information}

\item{hr_out}{an optional character vector describing variables of
interest for hazard ratio calculation}

\item{prevalence_data}{matrix of prevalences as class \code{paf_data}}

\item{group_vars}{names of columns to group by in PAF difference estimations}

\item{level}{width of confidence intervals, default 0.95}

\item{...}{extra parameters to be passed to \code{\link{est_matrix}}}
}
\value{
a list of potenitally interesting hazard ratio and PAF estimates
}
\description{
Calculate mortality PAF estimates, confidence intervals and groupwise
differences
}
